Consolidate HPC schedulers
