+++ src/meta_scheduler.py
@@ -8,7 +8,7 @@ from typing import Deque, Dict, List, Tuple, Union
 
 from .carbon_aware_scheduler import CarbonAwareScheduler
 from .rl_schedulers import RLCarbonScheduler
-from .hpc_base_scheduler import HPCBaseScheduler
+from .hpc_schedulers import HPCBaseScheduler
 
 SchedulerType = Union[
     CarbonAwareScheduler,
