+++ tests/test_rl_cost_scheduler.py
@@ -79,7 +79,7 @@ TelemetryLogger = _load('asi.telemetry', 'src/telemetry.py').TelemetryLogger
 _load('asi.hpc_schedulers', 'src/hpc_schedulers.py')
 rl_mod = _load('asi.rl_schedulers', 'src/rl_schedulers.py')
 RLCostScheduler = rl_mod.RLCostScheduler
-base_mod = _load('asi.hpc_base_scheduler', 'src/hpc_base_scheduler.py')
+base_mod = _load('asi.hpc_schedulers', 'src/hpc_schedulers.py')
 strat_mod = _load('asi.forecast_strategies', 'src/forecast_strategies.py')
 make_scheduler = base_mod.make_scheduler
